Break out tests for moved methods
